Finishing touches for Assingment 2 - Question 2

+++ headers/my-stack.h
@@ -16,7 +16,7 @@ public:
     int pop();
     bool isEmpty();
     int top();
-    double average();
+    int average();
 };
 
 #endif
